Moved process semver and parse under processerinput
