+++ src/processor/semver_test.go
@@ -20,9 +20,10 @@ func TestProcessSemverMainWithTagDefaults(t *testing.T) {
 	lib.CheckIfError(err)
 	_, err = r.CreateTag("v1.0.1-rc.123", commit.Hash, tagOptions)
 	lib.CheckIfError(err)
-	version, err := ProcessSemver(&ProcessorInput{
+	input := ProcessorInput{
 		RepoPath: dir,
-	})
+	}
+	version, err := input.ProcessSemver()
 	lib.CheckIfError(err)
 	expectedVersion := "1.0.1"
 	if *version != expectedVersion {
@@ -36,9 +37,10 @@ func TestProcessSemverNonMainWithDefaults(t *testing.T) {
 	t.Setenv("GITHUB_REF_NAME", "feature/123")
 	t.Setenv("GITHUB_RUN_ID", "abc")
 	dir, _, _, _ := bvtest.SetupRepo("main", t)
-	version, err := ProcessSemver(&ProcessorInput{
+	input := ProcessorInput{
 		RepoPath: dir,
-	})
+	}
+	version, err := input.ProcessSemver()
 	lib.CheckIfError(err)
 	expectedVersion := "0.0.1-rc.abc"
 	if *version != expectedVersion {
@@ -57,9 +59,10 @@ func TestProcessSemverNonMainWithExistingTagDefaults(t *testing.T) {
 	}
 	_, err := r.CreateTag("v1.0.0", commit.Hash, tagOptions)
 	lib.CheckIfError(err)
-	version, err := ProcessSemver(&ProcessorInput{
+	input := ProcessorInput{
 		RepoPath: dir,
-	})
+	}
+	version, err := input.ProcessSemver()
 	lib.CheckIfError(err)
 	expectedVersion := "1.0.1-rc.abc"
 	if *version != expectedVersion {
@@ -78,11 +81,12 @@ func TestProcessSemverNonMainWithExistingTagPatch(t *testing.T) {
 	}
 	_, err := r.CreateTag("v1.0.0", commit.Hash, tagOptions)
 	lib.CheckIfError(err)
-	version, err := ProcessSemver(&ProcessorInput{
+	input := ProcessorInput{
 		RepoPath:      dir,
 		IncrementType: "patch",
 		OfflineMode:   false,
-	})
+	}
+	version, err := input.ProcessSemver()
 	lib.CheckIfError(err)
 	expectedVersion := "1.0.1-rc.abc"
 	if *version != expectedVersion {
@@ -101,11 +105,12 @@ func TestProcessSemverNonMainWithExistingTagMinor(t *testing.T) {
 	}
 	_, err := r.CreateTag("v1.0.0", commit.Hash, tagOptions)
 	lib.CheckIfError(err)
-	version, err := ProcessSemver(&ProcessorInput{
+	input := ProcessorInput{
 		RepoPath:      dir,
 		IncrementType: "minor",
 		OfflineMode:   false,
-	})
+	}
+	version, err := input.ProcessSemver()
 	lib.CheckIfError(err)
 	expectedVersion := "1.1.0-rc.abc"
 	if *version != expectedVersion {
@@ -124,15 +129,95 @@ func TestProcessSemverNonMainWithExistingTagMajor(t *testing.T) {
 	}
 	_, err := r.CreateTag("v1.0.0", commit.Hash, tagOptions)
 	lib.CheckIfError(err)
-	version, err := ProcessSemver(&ProcessorInput{
+	input := ProcessorInput{
 		RepoPath:      dir,
 		IncrementType: "major",
 		OfflineMode:   false,
-	})
+	}
+	version, err := input.ProcessSemver()
 	lib.CheckIfError(err)
 	expectedVersion := "2.0.0-rc.abc"
 	if *version != expectedVersion {
 		t.Error(fmt.Sprintf("Expected %s but received %s", *version, expectedVersion))
 		t.Fail()
+	}
+}
+
+func TestGetIncrementTypeEmptyOffline(t *testing.T) {
+	input := ProcessorInput{
+		OfflineMode:   false,
+		IncrementType: "",
+	}
+	result, err := input.parseIncrementType()
+	if err != nil {
+		t.Fail()
+	}
+	if *result != "patch" {
+		t.Fail()
+	}
+}
+
+func TestGetIncrementTypePatchOffline(t *testing.T) {
+	incrementTypes := []string{"Patch", "patch"}
+	for _, incrementType := range incrementTypes {
+		input := ProcessorInput{
+			OfflineMode:   false,
+			IncrementType: incrementType,
+		}
+		result, err := input.parseIncrementType()
+		if err != nil {
+			t.Fail()
+		}
+		if *result != "patch" {
+			t.Fail()
+		}
+	}
+}
+
+func TestGetIncrementTypeMinorOffline(t *testing.T) {
+	incrementTypes := []string{"Minor", "minor"}
+	for _, incrementType := range incrementTypes {
+		input := ProcessorInput{
+			OfflineMode:   false,
+			IncrementType: incrementType,
+		}
+		result, err := input.parseIncrementType()
+		if err != nil {
+			t.Fail()
+		}
+		if *result != "minor" {
+			t.Fail()
+		}
+	}
+}
+
+func TestGetIncrementTypeMajorOffline(t *testing.T) {
+	incrementTypes := []string{"Major", "major"}
+	for _, incrementType := range incrementTypes {
+		input := ProcessorInput{
+			OfflineMode:   false,
+			IncrementType: incrementType,
+		}
+		result, err := input.parseIncrementType()
+		if err != nil {
+			t.Fail()
+		}
+		if *result != "major" {
+			t.Fail()
+		}
+	}
+}
+
+func TestGetIncrementTypeInvalidValueffline(t *testing.T) {
+	incrementTypes := []string{"asdasda"}
+	for _, incrementType := range incrementTypes {
+		input := ProcessorInput{
+			OfflineMode:   false,
+			IncrementType: incrementType,
+		}
+		result, err := input.parseIncrementType()
+		if result != nil && err == nil {
+			t.Fail()
+		}
 	}
 }
